+++ src/main/java/com/flytrap/venusplanner/api/auth_member/business/service/AuthMemberService.java
@@ -2,8 +2,8 @@ package com.flytrap.venusplanner.api.auth_member.business.service;
 
 import com.flytrap.venusplanner.api.member.infrastructure.repository.MemberRepository;
 import com.flytrap.venusplanner.api.member.domain.Member;
-import com.flytrap.venusplanner.api.auth_member.presentation.dto.LoginDto;
-import com.flytrap.venusplanner.global.auth.infrastructure.api.OAuthProvider;
+import com.flytrap.venusplanner.api.auth_member.presentation.dto.SignInDto;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.external.OAuthProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +17,7 @@ public class AuthMemberService {
     private final MemberRepository memberRepository;
 
     @Transactional
-    public Member authenticateAndFetchMember(LoginDto.Request request) {
+    public Member authenticateAndFetchMember(SignInDto.Request request) {
 
         var userResource = oAuthProvider.authenticateAndFetchUserResource(request.code());
         var authenticatedMember = memberRepository.findByOauthPk(userResource.oauthPk())
