+++ src/main/java/com/flytrap/venusplanner/api/auth_member/exception/GitHubOAuthRequestException.java
@@ -1,4 +1,4 @@
-package com.flytrap.venusplanner.global.auth.exception;
+package com.flytrap.venusplanner.api.auth_member.exception;
 
 public class GitHubOAuthRequestException extends RuntimeException {
     public GitHubOAuthRequestException(String message) {
