+++ src/main/java/com/flytrap/venusplanner/api/auth_member/infrastructure/external/GitHubOAuthProvider.java
@@ -1,11 +1,11 @@
-package com.flytrap.venusplanner.global.auth.infrastructure.api;
+package com.flytrap.venusplanner.api.auth_member.infrastructure.external;
 
 import com.flytrap.venusplanner.global.auth.config.OAuthWebClientConfig.GitHubOAuthFormDataBuilder;
-import com.flytrap.venusplanner.global.auth.exception.GitHubOAuthRequestException;
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.AccessTokenFromGitHub;
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.StandardizedUserResource;
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.UserEmailResourceFromGitHub;
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.UserResourceFromGitHub;
+import com.flytrap.venusplanner.api.auth_member.exception.GitHubOAuthRequestException;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.dto.AccessTokenFromGitHub;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.dto.StandardizedUserResource;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.dto.UserEmailResourceFromGitHub;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.dto.UserResourceFromGitHub;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
