+++ src/main/java/com/flytrap/venusplanner/api/member/domain/Member.java
@@ -1,22 +1,24 @@
 package com.flytrap.venusplanner.api.member.domain;
 
-import com.flytrap.venusplanner.global.auth.infrastructure.dto.StandardizedUserResource;
+import com.flytrap.venusplanner.api.auth_member.infrastructure.dto.StandardizedUserResource;
+import com.flytrap.venusplanner.global.entity.TimeAuditingBaseEntity;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
-import java.time.Instant;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
+@EntityListeners(AuditingEntityListener.class)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
+public class Member extends TimeAuditingBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -37,11 +39,8 @@ public class Member {
     @NotNull
     private String nickname;
 
-    @CreationTimestamp
-    private Instant createdTime;
-
     @Builder
-    public Member(String oauthPk, Long oauthPlatformId, String email, String profileImageUrl, String nickname) {
+    private Member(String oauthPk, Long oauthPlatformId, String email, String profileImageUrl, String nickname) {
         this.oauthPk = oauthPk;
         this.oauthPlatformId = oauthPlatformId;
         this.email = email;
